Move create_disallowed_map to clippy_config
