+++ clippy_config/Cargo.toml
@@ -2,10 +2,12 @@
 name = "clippy_config"
 version = "0.1.84"
 edition = "2021"
+publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clippy_utils = { path = "../clippy_utils" }
 itertools = "0.12"
 serde = { version = "1.0", features = ["derive"] }
 toml = "0.7.3"
