+++ clippy_config/src/lib.rs
@@ -14,6 +14,8 @@
 )]
 
 extern crate rustc_errors;
+extern crate rustc_hir;
+extern crate rustc_middle;
 extern crate rustc_session;
 extern crate rustc_span;
