+++ clippy_utils/Cargo.toml
@@ -2,10 +2,8 @@
 name = "clippy_utils"
 version = "0.1.84"
 edition = "2021"
-publish = false
 
 [dependencies]
-clippy_config = { path = "../clippy_config" }
 arrayvec = { version = "0.7", default-features = false }
 itertools = "0.12"
 # FIXME(f16_f128): remove when no longer needed for parsing
